Bug fixes: Issue 70 - Remember preference for Request [RawView|Tabbed View] Issue 103 - Rework/Refactor HttpBody.class to use a byte oriented buffer instead of StringBuffer Issue 107 - The last intercepted request/response remains in the Break window Issue 180 - Tabular view for GET request Issue 187 - Encoding issues Issue 197 - Decoder cannot process base64 input without line breaks Issue 237 - Bug: Problems with HTTP Panels Issue 260 - Exception when deleting ("Purge (from DB)") "History" tab entries Issue 264 - Exception in "Manual Request Editor"/"Resend" dialogues
Enhancements:
Issue 102 - Save raw response and request, and save body of response and request
Issue 242 - Export node / req/resp via right click

+++ src/org/zaproxy/zap/extension/params/ParamsTableModel.java
@@ -92,7 +92,7 @@ public class ParamsTableModel extends AbstractTableModel {
 		if (param.getValues().size() == 1) {
 			return 0;
 		}
-		return (int)(param.getValues().size() * 100 / param.getTimesUsed());
+		return (param.getValues().size() * 100 / param.getTimesUsed());
 	}
 
 	public HtmlParameterStats getHtmlParameterStatsAtRow(int row) {
